web3js: Split eth/sign.ts
